feat(auth): rework frontend api services for new best practices
First and foremost this commit moves calendar services into the calendar
module as despite the fact the api will always be there the api calls
shouldn't be considered core. The same goes for calendar models, which
have also been renamed to match auth api conventions

The calendar api services have been merged with the data services as
having them separate repeated a LOT of code. They have been reworked for
better type casting, also to match the auth api. This is in preparation
for proper logging and handling of errors.

A redundant feature from the backend has been removed as it was only
implemented in preparation for a feature that will never come.

Testing must be implemented as a next step